+++ lib/nanoc/base/repos/prefixed_data_source.rb
@@ -1,4 +1,4 @@
-module Nanoc
+module Nanoc::Int
   class PrefixedDataSource < Nanoc::DataSource
     def initialize(data_source, items_prefix, layouts_prefix)
       super({}, '/', '/', {})
